restructure the VS solution
Move everything that requires Entity Framework to the DataAccess project
and remove the Api dependency from Tests
