--- DOL.WHD.Section14c.Business/packages.config
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="EntityFramework" version="6.1.0" targetFramework="net452" />
-  <package id="Microsoft.AspNet.Identity.Core" version="2.2.1" targetFramework="net452" />
-  <package id="Microsoft.AspNet.Identity.EntityFramework" version="2.2.1" targetFramework="net452" />
-  <package id="zxcvbn-core" version="1.1.10" targetFramework="net452" />
-</packages>
