+++ DOL.WHD.Section14c.DataAccess/Identity/IdentityConfig.cs
@@ -1,15 +1,14 @@
 ﻿using System;
 using System.Configuration;
 using System.Threading.Tasks;
-using DOL.WHD.Section14c.Business.Validators;
-using DOL.WHD.Section14c.DataAccess;
+using DOL.WHD.Section14c.DataAccess.Validators;
 using DOL.WHD.Section14c.Domain.Models;
 using Microsoft.AspNet.Identity;
 using Microsoft.AspNet.Identity.EntityFramework;
 using Microsoft.AspNet.Identity.Owin;
 using Microsoft.Owin;
 
-namespace DOL.WHD.Section14c.Api
+namespace DOL.WHD.Section14c.DataAccess.Identity
 {
     // Configure the application user manager used in this application. UserManager is defined in ASP.NET Identity and is used by the application.
