+++ DOL.WHD.Section14c.DataAccess/Validators/Section14cPasswordValidator.cs
@@ -1,11 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.AspNet.Identity;
 
-namespace DOL.WHD.Section14c.Business.Validators
+namespace DOL.WHD.Section14c.DataAccess.Validators
 {
     public class Section14cPasswordValidator : PasswordValidator
     {
