+++ DOL.WHD.Section14c.DataAccess/packages.config
@@ -4,5 +4,13 @@
   <package id="EntityFramework6.Npgsql" version="3.1.1" targetFramework="net461" />
   <package id="Microsoft.AspNet.Identity.Core" version="2.2.1" targetFramework="net461" />
   <package id="Microsoft.AspNet.Identity.EntityFramework" version="2.2.1" targetFramework="net461" />
+  <package id="Microsoft.AspNet.Identity.Owin" version="2.2.1" targetFramework="net452" />
+  <package id="Microsoft.Owin" version="2.1.0" targetFramework="net452" />
+  <package id="Microsoft.Owin.Security" version="2.1.0" targetFramework="net452" />
+  <package id="Microsoft.Owin.Security.Cookies" version="2.1.0" targetFramework="net452" />
+  <package id="Microsoft.Owin.Security.OAuth" version="2.1.0" targetFramework="net452" />
+  <package id="Newtonsoft.Json" version="4.5.11" targetFramework="net452" />
   <package id="Npgsql" version="3.1.7" targetFramework="net461" />
+  <package id="Owin" version="1.0" targetFramework="net452" />
+  <package id="zxcvbn-core" version="1.1.10" targetFramework="net452" />
 </packages>
