+++ DOL.WHD.Section14c.Test/Business/Validators/Section14cUserValidatorTests.cs
@@ -3,8 +3,8 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
-using DOL.WHD.Section14c.Api;
-using DOL.WHD.Section14c.Business.Validators;
+using DOL.WHD.Section14c.DataAccess.Identity;
+using DOL.WHD.Section14c.DataAccess.Validators;
 using DOL.WHD.Section14c.Domain.Models;
 using Microsoft.AspNet.Identity.EntityFramework;
