+++ DOL.WHD.Section14c.Test/DOL.WHD.Section14c.Test.csproj
@@ -86,10 +86,6 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\DOL.WHD.Section14c.Api\DOL.WHD.Section14c.Api.csproj">
-      <Project>{7f947d49-de51-4b9e-95e0-afa3d5577c59}</Project>
-      <Name>DOL.WHD.Section14c.Api</Name>
-    </ProjectReference>
     <ProjectReference Include="..\DOL.WHD.Section14c.Business\DOL.WHD.Section14c.Business.csproj">
       <Project>{7cd1d7dd-ce24-4280-b8bd-0b8b9abb0ffc}</Project>
       <Name>DOL.WHD.Section14c.Business</Name>
